Rename commad.ts to command.js
