+++ src/cmds/playlist/getId.ts
@@ -1,6 +1,6 @@
 import { SlashCommandBuilder } from '@discordjs/builders';
 import { CommandInteraction } from 'discord.js';
-import { Command } from '../../commad.js';
+import { Command } from '../../command.js';
 import { createGuild, getPlayistId } from '../../db.js';
 
 export default {
