+++ src/cmds/playlist/link.ts
@@ -1,6 +1,6 @@
 import { SlashCommandBuilder } from '@discordjs/builders';
 import { CommandInteraction, GuildMember, MessageEmbed } from 'discord.js';
-import { Command } from '../../commad.js';
+import { Command } from '../../command.js';
 import { createGuild, linkPlaylist, playlistExists } from '../../db.js';
 
 export default {
